# api/app/docs/swagger.yaml
definitions:
...
  hrms_modules_salaryraise_internal_dto.Cycle:
    properties:
      createdAt:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Item'
        type: array
      periodEndDate:
        type: string
      periodStartDate:
        type: string
      status:
        type: string
      totalEmployees:
        type: integer
      totalRaiseAmount:
        type: number
      updatedAt:
        type: string
    type: object
  hrms_modules_salaryraise_internal_dto.Item:
    properties:
      currentSalary:
        type: number
      currentSsoWage:
        type: number
      employeeId:
        type: string
      employeeName:
        type: string
      employeeNumber:
        type: string
      id:
        type: string
      newSalary:
        type: number
      newSsoWage:
        type: number
      photoId:
        type: string
      raiseAmount:
        type: number
      raisePercent:
        type: number
      stats:
        $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Stats'
      tenureDays:
        type: integer
      updatedAt:
        type: string
    type: object
  hrms_modules_salaryraise_internal_dto.Meta:
    properties:
      currentPage:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
  hrms_modules_salaryraise_internal_dto.Stats:
    properties:
      lateMinutes:
        type: integer
      leaveDays:
        type: number
      leaveDoubleDays:
        type: number
      leaveHours:
        type: number
      otHours:
        type: number
    type: object
  hrms_modules_salaryraise_internal_feature_create.Command:
    properties:
      periodEndDate:
        type: string
      periodStartDate:
        type: string
    required:
    - periodEndDate
    - periodStartDate
    type: object
  hrms_modules_salaryraise_internal_feature_create.Response:
    properties:
      createdAt:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Item'
        type: array
      periodEndDate:
        type: string
      periodStartDate:
        type: string
      status:
        type: string
      totalEmployees:
        type: integer
      totalRaiseAmount:
        type: number
      updatedAt:
        type: string
    type: object
...
  hrms_modules_salaryraise_internal_feature_items_list.Response:
    properties:
      data:
        items:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_repository.Item'
        type: array
    type: object
...
  hrms_modules_salaryraise_internal_feature_list.Response:
    properties:
      data:
        items:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Cycle'
        type: array
      meta:
        $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Meta'
    type: object
...
  hrms_modules_salaryraise_internal_repository.Item:
    properties:
      currentSalary:
        type: number
      currentSsoWage:
        type: number
      cycleId:
        type: string
      employeeId:
        type: string
      employeeName:
        type: string
      employeeNumber:
        type: string
      id:
        type: string
      newSalary:
        type: number
      newSsoWage:
        type: number
      photoId:
        type: string
      raiseAmount:
        type: number
      raisePercent:
        type: number
      stats:
        $ref: '#/definitions/hrms_modules_salaryraise_internal_repository.Stats'
      tenureDays:
        type: integer
      updatedAt:
        type: string
    type: object
  hrms_modules_salaryraise_internal_repository.Stats:
    properties:
      lateMinutes:
        type: integer
      leaveDays:
        type: number
      leaveDoubleDays:
        type: number
      leaveHours:
        type: number
      otHours:
        type: number
    type: object
...
  hrms_modules_user_internal_feature_create.RequestBody:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  hrms_modules_user_internal_feature_create.User:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  salaryraise_internal_feature_items_update.Command:
    properties:
      id:
        type: string
      newSsoWage:
        type: number
      raiseAmount:
        type: number
      raisePercent:
        type: number
    type: object
  salaryraise_internal_feature_items_update.Response:
    properties:
      currentSalary:
        type: number
      currentSsoWage:
        type: number
      cycleId:
        type: string
      employeeId:
        type: string
      employeeName:
        type: string
      employeeNumber:
        type: string
      id:
        type: string
      newSalary:
        type: number
      newSsoWage:
        type: number
      photoId:
        type: string
      raiseAmount:
        type: number
      raisePercent:
        type: number
      stats:
        $ref: '#/definitions/hrms_modules_salaryraise_internal_repository.Stats'
      tenureDays:
        type: integer
      updatedAt:
        type: string
    type: object
...
  user_internal_feature_get.User:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  userbranch_internal_feature_getbranches.BranchAccess:
    properties:
      branchId:
        type: string
      code:
        type: string
      isDefault:
        type: boolean
      name:
        type: string
    type: object
  userbranch_internal_feature_listusers.CompanyUser:
    properties:
      createdAt:
        type: string
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  userbranch_internal_feature_setbranches.BranchAccess:
    properties:
      branchId:
        type: string
      code:
        type: string
      isDefault:
        type: boolean
      name:
        type: string
    type: object
  userbranch_internal_feature_setbranches.RequestBody:
    properties:
      branchIds:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: HR Payroll REST API
  title: HR Payroll API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      parameters:
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/userbranch_internal_feature_listusers.CompanyUser'
            type: array
      security:
      - BearerAuth: []
      summary: List company users
      tags:
      - User-Branch Access
    post:
      consumes:
      - application/json
      description: สร้างผู้ใช้งานใหม่ (Admin)
      parameters:
      - description: user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hrms_modules_user_internal_feature_create.RequestBody'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hrms_modules_user_internal_feature_create.User'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Conflict
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Admin Users
  /admin/users/{id}:
    delete:
      description: ลบผู้ใช้งาน (Admin)
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete user (soft delete)
      tags:
      - Admin Users
    get:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_internal_feature_get.User'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Admin Users
...
  /admin/users/{userId}/branches:
    get:
      parameters:
      - description: user ID
        in: path
        name: userId
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/userbranch_internal_feature_getbranches.BranchAccess'
            type: array
      security:
      - BearerAuth: []
      summary: Get user branch access
      tags:
      - User-Branch Access
    put:
      consumes:
      - application/json
      parameters:
      - description: user ID
        in: path
        name: userId
        required: true
        type: string
      - description: branches
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userbranch_internal_feature_setbranches.RequestBody'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/userbranch_internal_feature_setbranches.BranchAccess'
            type: array
      security:
      - BearerAuth: []
      summary: Set user branches
      tags:
      - User-Branch Access
...
  /salary-raise-cycles:
    get:
      description: ดึงรอบปรับเงินเดือน
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: pending|approved|rejected|all
        in: query
        name: status
        type: string
      - description: filter by periodStartDate year
        in: query
        name: year
        type: integer
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_salaryraise_internal_feature_list.Response'
      security:
      - BearerAuth: []
      summary: List salary raise cycles
      tags:
      - Salary Raise
    post:
      consumes:
      - application/json
      description: สร้างรอบปรับเงินเดือน (pending) มีได้ 1 รอบต่อสาขา
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_feature_create.Command'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hrms_modules_salaryraise_internal_feature_create.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Create salary raise cycle
      tags:
      - Salary Raise
...
  /salary-raise-cycles/{id}/items:
    get:
      parameters:
      - description: cycle id
        in: path
        name: id
        required: true
        type: string
      - description: search employee name
        in: query
        name: search
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_salaryraise_internal_feature_items_list.Response'
      security:
      - BearerAuth: []
      summary: List salary raise items
      tags:
      - Salary Raise
  /salary-raise-items/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: item id
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/salaryraise_internal_feature_items_update.Command'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/salaryraise_internal_feature_items_update.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update salary raise item
      tags:
      - Salary Raise
...
securityDefinitions:
  BearerAuth:
    description: Add "Bearer <access token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"